+++ maeumgagym-infrastructure/src/main/kotlin/com/info/maeumgagym/filter/config/SecurityFilterChainConfig.kt
@@ -7,7 +7,7 @@ import com.info.maeumgagym.error.filter.filterchain.ExceptionChainedFilterChainP
 import com.info.maeumgagym.error.repository.ExceptionRepository
 import com.info.maeumgagym.response.writer.DefaultHttpServletResponseWriter
 import com.info.maeumgagym.response.writer.ErrorLogHttpServletResponseWriter
-import com.info.maeumgagym.security.jwt.AuthenticationProvider
+import com.info.maeumgagym.security.authentication.AuthenticationProvider
 import com.info.maeumgagym.security.jwt.JwtFilter
 import com.info.maeumgagym.security.jwt.JwtResolver
 import com.info.maeumgagym.security.jwt.env.JwtProperties
