+++ maeumgagym-infrastructure/src/main/kotlin/com/info/maeumgagym/security/authentication/AuthenticationProvider.kt
@@ -1,0 +1,22 @@
+package com.info.maeumgagym.security.authentication
+
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
+
+/**
+ * [org.springframework.security.core.Authentication] 객체를 생성해주는 클래스
+ *
+ * @author Daybreak312
+ * @since 20-03-2024
+ */
+interface AuthenticationProvider {
+
+    /**
+     * 생성한 객체의 [org.springframework.security.core.Authentication.getPrincipal]에 User 정보를 담아 반환
+     */
+    fun getAuthentication(subject: String): UsernamePasswordAuthenticationToken
+
+    /**
+     * 생성한 객체의[org.springframework.security.core.Authentication.getPrincipal]을 null로 생성해 반환
+     */
+    fun getEmptyAuthentication(subject: String): UsernamePasswordAuthenticationToken
+}
