+++ maeumgagym-infrastructure/src/main/kotlin/com/info/maeumgagym/security/authentication/impl/AuthenticationProviderImpl.kt
@@ -1,7 +1,7 @@
-package com.info.maeumgagym.security.jwt.impl
+package com.info.maeumgagym.security.authentication.impl
 
 import com.info.maeumgagym.common.exception.CriticalException
-import com.info.maeumgagym.security.jwt.AuthenticationProvider
+import com.info.maeumgagym.security.authentication.AuthenticationProvider
 import com.info.maeumgagym.security.jwt.JwtFilter
 import com.info.maeumgagym.user.port.out.ReadUserPort
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
